requirements.txt
tencentcloud_sdk_python_common==3.0.1186
tencentcloud_sdk_python_hunyuan==3.0.1186
cprint==1.2.2
feedparser==6.0.11
lxml==4.9.4
PyYAML==6.0.1
Requests==2.31.0
schedule==1.2.1
urllib3==2.1.0
wcferry==39.2.4.0
websocket_client==1.5.1
chinese_calendar
bs4
pillow
lz4